Ignore internal Active Record migrations compatibility related options when suggesting a safe column type change
Before:

Changing the type of an existing column blocks reads and writes while the entire table is rewritten.
A safer approach can be accomplished in several steps:

1. Create a new column and keep column's data in sync:

  class InitializeCreateComments < ActiveRecord::Migration[7.1]
    def change
      initialize_column_type_change :users, :name, :float, _skip_validate_options: true
    end
  end

After:

Changing the type of an existing column blocks reads and writes while the entire table is rewritten.
A safer approach can be accomplished in several steps:

1. Create a new column and keep column's data in sync:

  class InitializeCreateComments < ActiveRecord::Migration[7.1]
    def change
      initialize_column_type_change :users, :name, :float
    end
  end